refactor: 스트라이크와 볼을 판단하는 기능 리팩토링
